Reduce complexity in LagrangianDynamics, MainWindow, and bench press renderer
- Extract LagrangianDynamics.__init__ (74 lines) into 4 focused helpers:
  _init_chain_geometry, _init_coupling_coefficients,
  _init_diagonal_mass_constants, _init_gravity_coefficients (#174)

- Decompose MainWindow (916 -> 618 lines) into 3 mixin modules:
  FileOperationsMixin, AnimationControlMixin, ComparisonMixin (#173)

- Break _draw_bench_press_frame (137 lines) into 4 composable steps:
  _draw_bench_and_body, _draw_bench_arm_chain,
  _draw_bench_barbell_and_trace, _draw_bench_title (#172)

All 265 tests pass. No workflow files modified.

Closes #172, closes #173, closes #174
